Add files via upload
UV Project 
UV stands for Ultraviolent light, and in this project, we are using UV C to prevent infection spread in hospital wards and theatre rooms. 
This project was funded by the Government of Uganda through the Ministry of Science Technology and Innovations in a bid to reduce the spread of hospital-acquired infections, antimicrobial resistance, and COVID-19. 
This being an embedded systems project, I used C++ to develop a code that would manage all process
First, we needed the system to be able to have a time before turning on the UV  C bulbs when switched on to allow the operate exit the room since UV C light is hazardous and that is why it can kill microbes. 
A beeping alarm was included to provide notice to the operator 
We also needed the system to detect motions using motion sensors and turn off the lights instantly if in case someone entered the room unknowingly 
There was a need for the system to operate both on battery and AC electricity supply. For this, the system needed to charge the battery using AC, and then the system draws power from the battery through the DC to DC converter.