Add files via upload
Low-Cost Touchless Handwashing Technology (TW-20 Kit)
Working with partners at Makerere University and Funding from Mak-RIF, Badaye Technologies, introduced the Green Low-Cost Touchless Handwashing Technology (TW-20 Kit) for Publicly Shared Spaces, This is the first of its kind mobile smart handwashing technology which uses a hybrid energy source of solar power, battery, and mains AC supply to automate the process of providing precise soap and rinsing water to the user from usually jerry cans. The audio user interface is incorporated to improve on usability with audio language tailored to the commonly spoken languages in the communities where we install the kits. 
To facilitate point-of-use user data mining, we connected TW-20 Kit to the TW-20 View; a custom-made data repository, curation, and analytics platform that receives handwashing user data remotely and reports it to the public. With TW-20 View, we are able to stream and interface using data from the handwashing kits in real time. Collected data is further analyzed automatically using back-end analytics algorithms to support public health programs in individual beneficiary organizations. Specifically, different sites are able to know the coping levels as regards hand washing on a daily and are able to introduce feasible handwashing interventions in their organizations.